Renamed files to lowercase, cleaned test_frontend.py
